Incremental commit - overnight changes
Tests in subdirectories of /modularodm/tests/ have been written or
reviewed by me. Tests in /modularodm/tests/ were written by Nan or
Melissa, and I have not fully reviewed them if they pass.

If those tests fail, I fix them and/or refactor them as necessary
and move them to a subdirectory as appropriate. If it is not immediately
apparent what they are testing, I skip them.