Move KeyExistsException to modulardom.exceptions
Exception was previously defined in modularodm.storage.base, and was not
a subclass of ModularOdmException. The definition was moved, and the
exception was changed to reflect the correct parent class. As it is
imported into the context where it was previously defined, existing code
that imports it from there will continue to work. New code should import
it from `modularodm.exceptions`.